Fixed typos, removed code to order dictionary before comparing
